+++ test-shell/src/test/java/scenario/EraseAndWriteAgingTest.java
@@ -1,8 +1,13 @@
+package scenario;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import shell.Processor;
+import shell.output.Output;
+import utils.RandomFactory;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -15,7 +20,7 @@ class EraseAndWriteAgingTest {
     ITestScenario testScenario;
 
     @Mock
-    RunCommand runCommand;
+    Processor processor;
 
     @Mock
     Output output;
@@ -25,13 +30,13 @@ class EraseAndWriteAgingTest {
 
     @BeforeEach
     void setUp() {
-        testScenario = new EraseAndWriteAging(runCommand, output, randomFactory);
+        testScenario = new EraseAndWriteAging(processor, output, randomFactory);
     }
 
     @Test
     void 정상적으로_모든_readCompare가_성공한_경우_return_true() {
         doReturn("0x99999999").when(randomFactory).getRandomHexValue();
-        doReturn(true).when(runCommand).execute(any());
+        doReturn(true).when(processor).execute(any());
         doReturn("LBA XX : 0x00000000").when(output).checkResult(anyString(), anyString());
 
         boolean actual = testScenario.run();
@@ -41,7 +46,7 @@ class EraseAndWriteAgingTest {
 
     @Test
     void value가_달라서_readCompare_실패한_경우_return_false() {
-        doReturn(true).when(runCommand).execute(any());
+        doReturn(true).when(processor).execute(any());
         doReturn("LBA XX : 0x12345678").when(output).checkResult(anyString(), anyString());
 
         boolean actual = testScenario.run();
@@ -51,7 +56,7 @@ class EraseAndWriteAgingTest {
 
     @Test
     void runCommand에서_false로_리턴한_경우_return_false() {
-        doReturn(false).when(runCommand).execute(any());
+        doReturn(false).when(processor).execute(any());
 
         boolean actual = testScenario.run();
